Reorganize tests and updates postboxes into upkeep page
