+++ js/upkeep.js
@@ -25,7 +25,6 @@ function validateEmail ($emailInput) {
   return result;
 }
 
-
 function drags(dragElement, resizeElement, container) {
 
   // Initialize the dragging event on mousedown.
@@ -39,9 +38,9 @@ function drags(dragElement, resizeElement, container) {
 
     // Get the initial position
     var dragWidth = dragElement.outerWidth(),
-        posX = dragElement.offset().left + dragWidth - startX,
-        containerOffset = container.offset().left,
-        containerWidth = container.outerWidth();
+      posX = dragElement.offset().left + dragWidth - startX,
+      containerOffset = container.offset().left,
+      containerWidth = container.outerWidth();
 
     // Set limits
     var minLeft = containerOffset + 10;
@@ -134,10 +133,10 @@ jQuery(document).ready(function($) {
   });
 
   jQuery.post(
-    seravo_updates_loc.ajaxurl, {
-      'action': 'seravo_ajax_updates',
+    seravo_upkeep_loc.ajaxurl, {
+      'action': 'seravo_ajax_upkeep',
       'section': 'seravo_default_config_file',
-      'nonce': seravo_updates_loc.ajax_nonce
+      'nonce': seravo_upkeep_loc.ajax_nonce
     },function(defaul_config_file) {
       if ( ! defaul_config_file ) {
         jQuery("#change-php-version-button").hide();
@@ -149,7 +148,7 @@ jQuery(document).ready(function($) {
 
   jQuery('#overwrite-config-files').change(function() {
     if (jQuery('#overwrite-config-files').is(':checked')) {
-        jQuery("#change-php-version-button").show();
+      jQuery("#change-php-version-button").show();
     } else {
       jQuery("#change-php-version-button").hide();
     }
@@ -160,21 +159,21 @@ jQuery(document).ready(function($) {
       jQuery(this).hide();
     });
     jQuery("#check-php-compatibility-status").fadeOut(400, function() {
-      jQuery(this).html('<img src="/wp-admin/images/spinner.gif" style="display:inline-block"> ' + seravo_updates_loc.compatibility_check_running).fadeIn(400);
+      jQuery(this).html('<img src="/wp-admin/images/spinner.gif" style="display:inline-block"> ' + seravo_upkeep_loc.compatibility_check_running).fadeIn(400);
     });
     checkPHPCompatibility();
   });
 
   function checkPHPCompatibility() {
     jQuery.post(
-      seravo_updates_loc.ajaxurl, {
-        'action': 'seravo_ajax_updates',
+      seravo_upkeep_loc.ajaxurl, {
+        'action': 'seravo_ajax_upkeep',
         'section': 'seravo_check_php_compatibility',
-        'nonce': seravo_updates_loc.ajax_nonce,
+        'nonce': seravo_upkeep_loc.ajax_nonce,
       },
       function(rawData) {
         if ( rawData.length == 0 ) {
-          jQuery('#check-php-compatibility-status').html(seravo_updates_loc.compatiblity_no_data);
+          jQuery('#check-php-compatibility-status').html(seravo_upkeep_loc.no_data);
         }
         var data = JSON.parse(rawData);
 
@@ -187,15 +186,15 @@ jQuery(document).ready(function($) {
             jQuery(this).show();
           });
           jQuery("#check-php-compatibility-status").fadeOut(400, function() {
-            jQuery(this).html('<div>' + seravo_updates_loc.compatibility_run_fail + '</div>').fadeIn(400);
+            jQuery(this).html('<div>' + seravo_upkeep_loc.compatibility_run_fail + '</div>').fadeIn(400);
           });
         } else if ( data['output'].length ) {
           jQuery("#check-php-compatibility-status").fadeOut(400, function() {
-            jQuery(this).html('<div style="color:red;font-weight:bold">' + data['output'] + seravo_updates_loc.compatibility_check_error + '</div>').fadeIn(400);
+            jQuery(this).html('<div style="color:red;font-weight:bold">' + data['output'] + seravo_upkeep_loc.compatibility_check_error + '</div>').fadeIn(400);
           });
         } else {
           jQuery("#check-php-compatibility-status").fadeOut(400, function() {
-            jQuery(this).html('<div style="color:green;font-weight:bold">' + seravo_updates_loc.compatibility_check_clear + '</div>').fadeIn(400);
+            jQuery(this).html('<div style="color:green;font-weight:bold">' + seravo_upkeep_loc.compatibility_check_clear + '</div>').fadeIn(400);
           });
         }
       }
@@ -211,11 +210,11 @@ jQuery(document).ready(function($) {
     jQuery("#overwrite-config-files-span").hide();
 
     jQuery.post(
-      seravo_updates_loc.ajaxurl, {
-        'action': 'seravo_ajax_updates',
+      seravo_upkeep_loc.ajaxurl, {
+        'action': 'seravo_ajax_upkeep',
         'section': 'check_php_config_files',
-        'nonce': seravo_updates_loc.ajax_nonce
-    });
+        'nonce': seravo_upkeep_loc.ajax_nonce
+      });
 
     jQuery("#seravo-php-version").fadeOut(400, function() {
       $(this).hide();
@@ -230,44 +229,44 @@ jQuery(document).ready(function($) {
     var php_version = $('[name=php-version]:checked').val();
 
     jQuery.post(
-      seravo_updates_loc.ajaxurl, {
-        'action': 'seravo_ajax_updates',
+      seravo_upkeep_loc.ajaxurl, {
+        'action': 'seravo_ajax_upkeep',
         'section': 'seravo_change_php_version',
-        'nonce': seravo_updates_loc.ajax_nonce,
+        'nonce': seravo_upkeep_loc.ajax_nonce,
         'version': php_version
-    });
+      });
     setTimeout(function() {
       jQuery.post(
-        seravo_updates_loc.ajaxurl, {
-          'action': 'seravo_ajax_updates',
+        seravo_upkeep_loc.ajaxurl, {
+          'action': 'seravo_ajax_upkeep',
           'section': 'seravo_php_check_version',
-          'nonce': seravo_updates_loc.ajax_nonce,
+          'nonce': seravo_upkeep_loc.ajax_nonce,
           'version': php_version
-      }, function(success) {
-        jQuery("#change-php-version-status").fadeOut(400, function() {
-          if (success) {
-            jQuery("#activated-line").fadeIn(400, function() {
-              jQuery(this).show();
-            });
-          } else {
-            jQuery("#activation-failed-line").fadeIn(400, function() {
-              jQuery(this).show();
-            });
-          }
-        });
+        }, function(success) {
+          jQuery("#change-php-version-status").fadeOut(400, function() {
+            if (success) {
+              jQuery("#activated-line").fadeIn(400, function() {
+                jQuery(this).show();
+              });
+            } else {
+              jQuery("#activation-failed-line").fadeIn(400, function() {
+                jQuery(this).show();
+              });
+            }
+          });
 
-        jQuery("#seravo-php-version").fadeIn(400, function() {
-          $(this).show();
+          jQuery("#seravo-php-version").fadeIn(400, function() {
+            $(this).show();
+          });
         });
-      });
     }, 5000);
   }
 
   jQuery.post(
-    seravo_updates_loc.ajaxurl, {
-      'action': 'seravo_ajax_updates',
+    seravo_upkeep_loc.ajaxurl, {
+      'action': 'seravo_ajax_upkeep',
       'section': 'seravo_plugin_version_check',
-      'nonce': seravo_updates_loc.ajax_nonce
+      'nonce': seravo_upkeep_loc.ajax_nonce
     }, function(is_uptodate_version) {
       if (is_uptodate_version) {
         $('#uptodate_seravo_plugin_version').show();
@@ -288,29 +287,103 @@ jQuery(document).ready(function($) {
 
   function update_seravo_plugin() {
     jQuery.post(
-      seravo_updates_loc.ajaxurl, {
-        'action': 'seravo_ajax_updates',
+      seravo_upkeep_loc.ajaxurl, {
+        'action': 'seravo_ajax_upkeep',
         'section': 'seravo_plugin_version_update',
-        'nonce': seravo_updates_loc.ajax_nonce
+        'nonce': seravo_upkeep_loc.ajax_nonce
       });
-      setTimeout(function() {
-        jQuery("#update_seravo_plugin_status").fadeOut(400, function() {
-          jQuery(this).hide();
-        });
-        jQuery("#seravo_plugin_updated").show();
-      }, 5000);
+    setTimeout(function() {
+      jQuery("#update_seravo_plugin_status").fadeOut(400, function() {
+        jQuery(this).hide();
+      });
+      jQuery("#seravo_plugin_updated").show();
+    }, 5000);
   }
 });
 
 jQuery(window).load(function(){
   jQuery('.ba-slider').each(function(){
-      var cur = jQuery(this);
-      // Adjust the slider
-      var width = cur.width() + 'px';
-      cur.find('.ba-resize img').css('width', width);
-      // Bind dragging events
-      drags(cur.find('.ba-handle'), cur.find('.ba-resize'), cur);
+    var cur = jQuery(this);
+    // Adjust the slider
+    var width = cur.width() + 'px';
+    cur.find('.ba-resize img').css('width', width);
+    // Bind dragging events
+    drags(cur.find('.ba-handle'), cur.find('.ba-resize'), cur);
   });
+
+  function seravo_load_test_report() {
+    console.log('load test report');
+    jQuery.post(
+      seravo_upkeep_loc.ajaxurl,
+      { 'action': 'seravo_ajax_upkeep',
+        'section': 'seravo_tests',
+        'nonce': seravo_upkeep_loc.ajax_nonce },
+      function(rawData) {
+        console.log(rawData);
+        if ( rawData.length == 0 ) {
+          jQuery('#seravo_tests').html(seravo_upkeep_loc.no_data);
+        }
+        jQuery('#seravo_tests_status').fadeOut('slow', function() {
+          var data = JSON.parse(rawData);
+          var data_joined = data['test_result'].join("\n");
+
+          if ( data['exit_code'] === 0 ) {
+            // No failures, if the return value from wp-test was 0
+            jQuery(this).html(seravo_upkeep_loc.test_success).fadeIn('slow');
+            jQuery('.seravo-test-result-wrapper').css('border-left', 'solid 0.5em #038103');
+          } else {
+            // At least 1 failure, if the return value was non-zero
+            jQuery(this).html(seravo_upkeep_loc.test_fail).fadeIn('slow');
+            jQuery('.seravo-test-result-wrapper').css('border-left', 'solid 0.5em #e74c3c');
+          }
+          // Display the retrieved data and re-enable the run tests button
+          jQuery('#seravo_tests').append(data_joined);
+          jQuery('#run-wp-tests').prop('disabled', false);
+
+          jQuery(this).fadeIn('slow', function() {
+            jQuery('#seravo_test_show_more_wrapper').fadeIn('slow');
+          });
+        });
+      }
+    ).fail(function() {
+      jQuery('#seravo_tests_status').html(seravo_upkeep_loc.run_fail);
+      jQuery('#run-wp-tests').prop('disabled', false);
+    });
+  }
+
+  jQuery('#run-wp-tests').click(function() {
+    jQuery('#seravo_tests').html('');
+    jQuery('.seravo-test-result-wrapper').css('border-left', 'solid 0.5em #e8ba1b');
+    jQuery('#seravo_test_show_more_wrapper').hide();
+
+    if ( jQuery('#seravo_arrow_show_more').hasClass('dashicons-arrow-up-alt2') ) {
+      jQuery('.seravo-test-result').hide(function() {
+        jQuery('#seravo_arrow_show_more').removeClass('dashicons-arrow-up-alt2').addClass('dashicons-arrow-down-alt2');
+      });
+    }
+
+    jQuery('#seravo_tests_status').fadeOut(400, function() {
+      jQuery(this).html('<img src="/wp-admin/images/spinner.gif" style="display:inline-block"> ' + seravo_upkeep_loc.running_tests).fadeIn(400);
+    });
+
+    jQuery(this).prop('disabled', true);
+    seravo_load_test_report();
+  });
+
+  jQuery('#seravo_test_show_more').click(function(event) {
+    event.preventDefault();
+
+    if ( jQuery('#seravo_arrow_show_more').hasClass('dashicons-arrow-down-alt2') ) {
+      jQuery('.seravo-test-result').slideDown('fast', function() {
+        jQuery('#seravo_arrow_show_more').removeClass('dashicons-arrow-down-alt2').addClass('dashicons-arrow-up-alt2');
+      });
+    } else if ( jQuery('#seravo_arrow_show_more').hasClass('dashicons-arrow-up-alt2') ) {
+      jQuery('.seravo-test-result').hide(function() {
+        jQuery('#seravo_arrow_show_more').removeClass('dashicons-arrow-up-alt2').addClass('dashicons-arrow-down-alt2');
+      });
+    }
+  });
+
 });
 
 // Update sliders on resize.
