+++ modules/wp-cli.php
@@ -23,8 +23,8 @@ class Seravo_WP_CLI extends WP_CLI_Command {
    */
   public function updates( $args, $assoc_args ) {
 
-    require_once dirname(__FILE__) . '/../modules/updates.php';
-    $site_info = Seravo\Updates::seravo_admin_get_site_info();
+    require_once dirname(__FILE__) . '/../modules/upkeep.php';
+    $site_info = Seravo\Upkeep::seravo_admin_get_site_info();
 
     if ( $site_info['seravo_updates'] === true ) {
       WP_CLI::success('Seravo Updates: enabled');
