--- style/tests.css
@@ -1,60 +1,0 @@
-.seravo-test-result-wrapper, .seravo-test-legacy-result-wrapper {
-  -moz-transition: border 1s ease-in;
-  -o-transition: border 1s ease-in;
-  -webkit-transition: border 1s ease-in;
-  background-color: white;
-  border-left: solid 0.5em #e8ba1b;
-  box-shadow: 1px 1px 1px 1px rgba(0,0,0,0.12);
-  display: block;
-  margin-top: 2em;
-  max-width: 55em;
-  min-height: 5em;
-  overflow: hidden;
-  transition: border 1s ease-in;
-}
-
-.seravo-test-result, .seravo-test-legacy-result {
-  box-shadow: 0 1px 1px 1px rgba(0,0,0,0.2);
-  display: none;
-  margin: 1em;
-  max-height: 25em;
-  overflow: auto;
-  padding: 1em;
-}
-
-.seravo-test-status, .seravo-test-legacy-status {
-  font-size: 15px;
-  font-weight: bold;
-  margin-top: 1.6em;
-  margin-left: 0.5em;
-  padding-bottom: 7px;
-  text-align: center;
-}
-
-/* @TODO: Use classes here, not ID selectors, as we can have many similar elements */
-#seravo_test_show_more_wrapper {
-  -ms-transform: translateY(-50%);
-  -webkit-transform: translateY(-50%);
-  float: right;
-  height: 1em;
-  margin-right: 0.5em;
-  position: relative;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-/* @TODO: Use classes here, not ID selectors, as we can have many similar elements */
-#seravo_test_show_more,
-#seravo_test_show_more:hover,
-#seravo_test_show_more:active,
-#seravo_test_show_more:focus {
-  box-shadow: none;
-  color: inherit;
-  text-decoration: none;
-}
-
-/* @TODO: Use classes here, not ID selectors, as we can have many similar elements */
-#seravo_tests_loading {
-  margin: 0 auto;
-  overflow: hidden;
-}
